+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - #2752 Add modified index to catalog
+- #2753 Enable full view reload for specific item transitions in listings
 - #2751 Apply worksheet template only to analyses of selected samples
 - #2748 Add extra check to make sure a result is a true string before cgi.replace call
 - #2745 Fix wrong labels for attachment "Render in Report" checkboxes